Move add domain task from free trial to ./task-headers
