+++ src/free-trial/fills/index.ts
@@ -1,4 +1,3 @@
 export * from './task-headers/payments';
 export * from './task-headers/woocommerce-payments';
-export * from './task-headers/add-domain';
 export * from './woocommerce-payments';
