+++ src/task-headers/assets/images/add-domain-illustration.png
